[IRGen] Include space for conformance requirements in witness tables.
This introduces a separate type for storing information about
conformances, since they aren't keyed off a Decl like the other witness
table entries. These are stored separately at this point of the
compiler, but they're all still indexed as a flat array.

The actual entries are null pointers for now.