WIP.
This is broken for

import Foundation
@unsafe_no_objc_tagged_pointer @objc
protocol NSCoreBlah {
    var count: Int { get }
}

because the signature flattening code in getLoweredFormalTypes has
changed behaviour, meaning the abstraction pattern ends up as
Type(T_0_0), which gives Opaque when indexed like a tuple for mapping
down to the relevant element. Probably needs to be conditionalized on
having one element or the function input accessors turned into an
ArrayRef<AbstractionPattern> to match params or something.